[AR-13] Move Login section and its components to a separate file
